# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/summarize_url": {
      "post": {
        "tags": [
          "summarize"
        ],
        "summary": "Summarize Route",
        "operationId": "summarize_route_v1_summarize_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/summarize_pdf_or_docx": {
      "post": {
        "tags": [
          "summarize"
        ],
        "summary": "Summarize Route",
        "operationId": "summarize_route_v1_summarize_pdf_or_docx_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/load": {
      "post": {
        "tags": [
          "load document"
        ],
        "summary": "Load Doc Route",
        "operationId": "load_doc_route_v1_load_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/uploadImage": {
      "post": {
        "tags": [
          "load document"
        ],
        "summary": "Upload Image",
        "operationId": "upload_image_v1_uploadImage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/uploadImageBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/rag/ingest": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Ingest Route",
        "description": "Use this api to get information from reference database",
        "operationId": "ingest_route_v1_rag_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/rag/query": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Query Route",
        "operationId": "query_route_v1_rag_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/rag/delete": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Delete Route",
        "operationId": "delete_route_v1_rag_delete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rag/get_chunk/{chunk_id}": {
      "get": {
        "tags": [
          "rag"
        ],
        "summary": "Get Chunk",
        "operationId": "get_chunk_v1_rag_get_chunk__chunk_id__get",
        "parameters": [
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestRequest": {
        "properties": {
          "file_url": {
            "type": "string",
            "title": "File Url",
            "default": "unknown"
          },
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name",
            "default": "unknown"
          },
          "embeddingsProvider": {
            "type": "string",
            "title": "Embeddingsprovider",
            "default": "unknown"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "unknown"
          }
        },
        "type": "object",
        "title": "IngestRequest"
      },
...
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "the information to be looked for"
          },
          "source_count": {
            "type": "integer",
            "title": "Source Count",
            "default": 5
          },
          "file_ids": {
            "items": {},
            "type": "array",
            "title": "File Ids",
            "description": "the ids of the files to be queried",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest"
      },
...
      "SummarizeRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "file_url",
          "model"
        ],
        "title": "SummarizeRequest",
        "examples": [
          {
            "file_url": "https://www.fhi360.org/sites/default/files/media/documents/cfpb_building_block_activities_what-is-insurance_handout.pdf",
            "model": "gpt-3.5-turbo-instruct",
            "prompt": "Write a concise summary of the following:{text}CONCISE SUMMARY:"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "uploadImageBody": {
        "properties": {
          "data_url": {
            "type": "string",
            "title": "Data Url",
            "description": "The data URL of the image"
          }
        },
        "type": "object",
        "required": [
          "data_url"
        ],
        "title": "uploadImageBody"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}